+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- fix: Change SentryScope setTagValue to NSString #524
+
 ## 5.0.5
 
 - feat: Add remove methods for SentryScope #529