+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - fix: Avoid Implicit conversion in SentrySession #540
+- fix: Change SentryScope setTagValue to NSString #524
 
 ## 5.0.5