+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- fix: Remove MemoryWarningIntegration #537
 - fix: Change SentryScope setTagValue to NSString #524
 
 ## 5.0.5